+++ src/Resources/Resources/ChangeLog.md
@@ -19,7 +19,7 @@
 -->
 
 ## Upcoming Release
-* Added `Get-AzResourceManagementPrivateLink`,`Get-AzResourceManagementPrivateLinks`,`New-AzResourceManagementPrivateLink`,            `Remove-AzResourceManagementPrivateLink`, `Get-AzResourceManagementPrivateLinkAssociations` and          `Remove-AzResourceManagementPrivateLinkAssociation` cmdlets
+* Added `Get-AzResourceManagementPrivateLink`, `Remove-AzResourceManagementPrivateLink`, `Get-AzResourceManagementPrivateLinkAssociation` and  `Remove-AzResourceManagementPrivateLinkAssociation` cmdlets
 
 ## Version 5.4.0
 * Fixed keycredential key format, from base64url to byte [#17131]